refactor(utils): import utility types from utils file
